Replace old JSON schema exporter (+tests)
